Pull Lift upload into it's own controller
